chrome_dino/requirements.txt
pyautogui
mss
opencv-python
pillow
numpy